requirements.txt
absl-py
cffi
click
clipboard
cloudpickle
colorama
cryptography
dask
easyocr
et-xmlfile
filelock
fsspec
grpcio
imageio
jinja2
keyboard
lazy_loader
locket
markdown
MarkupSafe
mouseinfo
mpmath
networkx
ninja
numpy
opencv-python-headless
openpyxl
packaging
partd
pillow
protobuf
pyautogui
pyclipper
pycparser
pygetwindow
pymsgbox
pymupdf
pyperclip
PyQt6
PyQt6-Qt6
PyQt6-sip
PyQt6-WebEngine
PyQt6-WebEngine-Qt6
pyrect
pyscreeze
pytesseract
python-bidi
python-dateutil
pytweening
PyYAML
scikit-image
scipy
setuptools # Nur einmal auflisten!
shapely
six
sympy # Nur einmal auflisten!
tensorboard
tensorboard-data-server
tifffile
toolz
torch # Nur einmal auflisten!
torchvision # Nur einmal auflisten!
typing-extensions
werkzeug